# BackendAPI/WebAPI/WebAPI/bin/Release/net7.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "WebAPI, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
    "/api/Account/Login": {
      "post": {
        "tags": [
          "Account"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/LoginReqDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Account/ResetPassword": {
      "patch": {
        "tags": [
          "Account"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/FormResetDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Account/Register": {
      "post": {
        "tags": [
          "Account"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/LoginReqDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Account/ListAccount": {
      "get": {
        "tags": [
          "Account"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Account/DeleteAccount": {
      "delete": {
        "tags": [
          "Account"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "name",
            "type": "string",
            "collectionFormat": "multi"
          },
          {
            "in": "query",
            "name": "nameToDelete",
            "type": "string",
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Citys/GetCity": {
      "get": {
        "tags": [
          "Citys"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Citys/PostCity": {
      "post": {
        "tags": [
          "Citys"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CityDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Citys/Update/{id}": {
      "put": {
        "tags": [
          "Citys"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CityDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Citys/DeleteCity/{id}": {
      "delete": {
        "tags": [
          "Citys"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/FurnishingType/List": {
      "get": {
        "tags": [
          "FurnishingType"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/FurnishingType/Details/{id}": {
      "get": {
        "tags": [
          "FurnishingType"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/FurnishingType/Delete/{id}": {
      "delete": {
        "tags": [
          "FurnishingType"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/FurnishingType/AddFurnishing": {
      "post": {
        "tags": [
          "FurnishingType"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/FurnishingType"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Property/list/{SellRent}": {
      "get": {
        "tags": [
          "Property"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "SellRent",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Property/TopView": {
      "get": {
        "tags": [
          "Property"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Property/View/{PropId}": {
      "patch": {
        "tags": [
          "Property"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "PropId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Property/Details/{id}": {
      "get": {
        "tags": [
          "Property"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Property/DetailsUpdate/{id}": {
      "get": {
        "tags": [
          "Property"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Property/MyProperty/{UserName}": {
      "get": {
        "tags": [
          "Property"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "UserName",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Property/Add": {
      "post": {
        "tags": [
          "Property"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/PropertyDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Property/Add/photo/{propid}": {
      "post": {
        "tags": [
          "Property"
        ],
        "consumes": [
          "multipart/form-data"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "propid",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "formData",
            "name": "file",
            "type": "file"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/PhotoDTO"
            }
          }
        }
      }
    },
    "/api/Property/set-primary-photo/{propId}/{photoPublicId}": {
      "post": {
        "tags": [
          "Property"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "propId",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "path",
            "name": "photoPublicId",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Property/delete-photo/{propId}/{photoPublicId}": {
      "delete": {
        "tags": [
          "Property"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "propId",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "path",
            "name": "photoPublicId",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Property/Update/{id}": {
      "put": {
        "tags": [
          "Property"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/PropertyDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/PropertyType/List": {
      "get": {
        "tags": [
          "PropertyType"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/PropertyType/Details/{id}": {
      "get": {
        "tags": [
          "PropertyType"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/PropertyType/Delete/{id}": {
      "delete": {
        "tags": [
          "PropertyType"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/PropertyType/AddCPropertyAsync": {
      "post": {
        "tags": [
          "PropertyType"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/PropertyType"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "definitions": {
    "CityDTO": {
      "required": [
        "country",
        "name"
      ],
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "name": {
          "maxLength": 50,
          "minLength": 2,
          "pattern": ".*[a-zA-Z]+.*",
          "type": "string"
        },
        "country": {
          "minLength": 1,
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "FormResetDto": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "oldPassword": {
          "type": "string"
        },
        "newPassword": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "FurnishingType": {
      "required": [
        "name"
      ],
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "lastUpdatedOn": {
          "format": "date-time",
          "type": "string"
        },
        "lastUpdatedBy": {
          "format": "int32",
          "type": "integer"
        },
        "name": {
          "minLength": 1,
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "LoginReqDto": {
      "required": [
        "name",
        "password"
      ],
      "type": "object",
      "properties": {
        "name": {
          "minLength": 1,
          "type": "string"
        },
        "password": {
          "minLength": 1,
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "phonenumber": {
          "type": "string"
        },
        "otherContactInformation": {
          "maxLength": 300,
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "PhotoDTO": {
      "type": "object",
      "properties": {
        "imageUrl": {
          "type": "string"
        },
        "publicId": {
          "type": "string"
        },
        "isPrimary": {
          "type": "boolean"
        }
      },
      "additionalProperties": false
    },
    "PropertyDTO": {
      "type": "object",
      "properties": {
        "sellRent": {
          "format": "int32",
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "propertyTypeId": {
          "format": "int32",
          "type": "integer"
        },
        "furnishingTypeId": {
          "format": "int32",
          "type": "integer"
        },
        "price": {
          "format": "int32",
          "type": "integer"
        },
        "bhk": {
          "format": "int32",
          "type": "integer"
        },
        "builtArea": {
          "format": "int32",
          "type": "integer"
        },
        "cityId": {
          "format": "int32",
          "type": "integer"
        },
        "readyToMove": {
          "type": "boolean"
        },
        "carpetArea": {
          "format": "int32",
          "type": "integer"
        },
        "address": {
          "type": "string"
        },
        "address2": {
          "type": "string"
        },
        "floorNo": {
          "format": "int32",
          "type": "integer"
        },
        "totalFloors": {
          "format": "int32",
          "type": "integer"
        },
        "mainEntrance": {
          "type": "string"
        },
        "security": {
          "format": "int32",
          "type": "integer"
        },
        "gated": {
          "type": "boolean"
        },
        "maintenance": {
          "format": "int32",
          "type": "integer"
        },
        "estPossessionOn": {
          "format": "date-time",
          "type": "string"
        },
        "age": {
          "format": "int32",
          "type": "integer"
        },
        "description": {
          "type": "string"
        },
        "view": {
          "format": "int32",
          "type": "integer"
        }
      },
      "additionalProperties": false
    },
    "PropertyType": {
      "required": [
        "name"
      ],
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "lastUpdatedOn": {
          "format": "date-time",
          "type": "string"
        },
        "lastUpdatedBy": {
          "format": "int32",
          "type": "integer"
        },
        "name": {
          "minLength": 1,
          "type": "string"
        }
      },
      "additionalProperties": false
    }
  }
}